moving the focus element from app comp to the register component adding ngAfterViewInit() to resetFocus
