+++ src/register/components/check-your-answers/check-your-answers.component.ts
@@ -1,4 +1,4 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output, AfterViewInit} from '@angular/core';
 import {FormDataValuesModel} from '../../models/form-data-values.model';
 
 /**
@@ -10,7 +10,7 @@ import {FormDataValuesModel} from '../../models/form-data-values.model';
   selector: 'app-check-your-answers',
   templateUrl: './check-your-answers.component.html',
 })
-export class CheckYourAnswersComponent {
+export class CheckYourAnswersComponent implements AfterViewInit {
 
   constructor() {}
 
@@ -19,6 +19,14 @@ export class CheckYourAnswersComponent {
   @Output() submit = new EventEmitter();
   @Input() set fromValues(values) {
     this.formDataValues = values;
+  }
+
+  ngAfterViewInit() {
+    const focusElement = document.getElementsByTagName('h1')[0];
+    if (focusElement) {
+      focusElement.setAttribute('tabindex', '-1');
+      focusElement.focus();
+    }
   }
 
   onSubmitData() {
